update v1.5.2  2017-04-26 同步EasyUI v1.5.2

+++ plugins/jquery.parser.js
@@ -1,7 +1,7 @@
 /**
- * jQuery EasyUI 1.5.1
+ * jQuery EasyUI 1.5.2
  * 
- * Copyright (c) 2009-2016 www.jeasyui.com. All rights reserved.
+ * Copyright (c) 2009-2017 www.jeasyui.com. All rights reserved.
  *
  * Licensed under the freeware license: http://www.jeasyui.com/license_freeware.php
  * To use it on other terms please contact us: info@jeasyui.com
@@ -101,7 +101,7 @@ _13=_13||0;
 var v=$.trim(String(_11||""));
 var _14=v.substr(v.length-1,1);
 if(_14=="%"){
-v=parseInt(v.substr(0,v.length-1));
+v=parseFloat(v.substr(0,v.length-1));
 if(_10.toLowerCase().indexOf("width")>=0){
 v=Math.floor((_12.width()-_13)*v/100);
 }else{
