+++ src/jquery.parser.js
@@ -1,7 +1,7 @@
 /**
- * jQuery EasyUI 1.5.1
+ * jQuery EasyUI 1.5.2
  * 
- * Copyright (c) 2009-2016 www.jeasyui.com. All rights reserved.
+ * Copyright (c) 2009-2017 www.jeasyui.com. All rights reserved.
  *
  * Licensed under the freeware license: http://www.jeasyui.com/license_freeware.php
  * To use it on other terms please contact us: info@jeasyui.com
@@ -127,7 +127,7 @@
 			var v = $.trim(String(value||''));
 			var endchar = v.substr(v.length-1, 1);
 			if (endchar == '%'){
-				v = parseInt(v.substr(0, v.length-1));
+				v = parseFloat(v.substr(0, v.length-1));
 				if (property.toLowerCase().indexOf('width') >= 0){
 					v = Math.floor((parent.width()-delta) * v / 100.0);
 				} else {
